new symengine robot and tests
