+++ src/giskardpy/__init__.py
@@ -1,16 +1,16 @@
-# DEBUG = 0
-# NORMAL = 1
-# ERROR = 2
-# # TODO figure out how to do this print level shit properly
-# PRINT_LEVEL = NORMAL
+DEBUG = 0
+NORMAL = 1
+ERROR = 2
+# TODO figure out how to do this print level shit properly
+PRINT_LEVEL = NORMAL
 
 
-from giskardpy.pybullet_world import PyBulletWorldObj
 
 # BACKEND = None
 # BACKEND = 'cse'
 # BACKEND = 'numpy'
 # BACKEND = 'cython'
+from giskardpy.pybullet_world_object import PyBulletWorldObj
 
 BACKEND = 'llvm'
 # BACKEND = 'lambda'
