+++ src/giskardpy/data_types.py
@@ -11,74 +11,6 @@ class SingleJointState(object):
     def __str__(self):
         return u'{}: {}, {}, {}'.format(self.name, self.position, self.velocity, self.effort)
 
-
-# class MultiJointState(object):
-#     def __init__(self):
-#         self._states = OrderedDict()
-#
-#     def get(self, name):
-#         return self._states[name]
-#
-#     def set(self, state):
-#         self._states[state.name] = state
-#
-#     def keys(self):
-#         return self._states.keys()
-#
-#     def values(self):
-#         return self._states.values()
-#
-#     def items(self):
-#         return self._states.items()
-
-
-# class Point(object):
-#     def __init__(self, x=0.0, y=0.0, z=0.0):
-#         self.x = x
-#         self.y = y
-#         self.z = z
-#
-#     def __repr__(self):
-#         return u'{}(x={}, y={}, z={})'.format(self.__class__.__name__, self.x, self.y, self.z)
-#
-#
-# class Quaternion(object):
-#     def __init__(self, x=0.0, y=0.0, z=0.0, w=1.0):
-#         self.x = x
-#         self.y = y
-#         self.z = z
-#         self.w = w
-#
-#     def __repr__(self):
-#         return u'{}(x={}, y={}, z={}, w={})'.format(self.__class__.__name__, self.x, self.y, self.z, self.w)
-#
-#
-# class Transform(object):
-#     def __init__(self, translation=Point(), rotation=Quaternion()):
-#         self.translation = translation
-#         self.rotation = rotation
-#
-#     def __repr__(self):
-#         return u'{}(translation={}, rotation={})'.format(self.__class__.__name__, self.translation, self.rotation)
-
-
-# class Stamped(object):
-#     def __init__(self, reference_name=''):
-#         self.reference_name = reference_name
-
-
-# class PointStamped(Point, Stamped):
-#     pass
-
-
-# class QuaternionStamped(Quaternion, Stamped):
-#     pass
-
-
-# class TransformStamped(Transform, Stamped):
-#     pass
-
-
 class Trajectory(object):
     def __init__(self):
         self._points = OrderedDict()
